rename RootOptionsManager -> OptionsManager; OptionsManager (old) -> OptionsGroup; move OptionsManager stuff to options_manager.py
